% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\docType{methods}
\name{annotation_matches}
\alias{annotation_matches}
\alias{annotation_matches,SummarizedExperiment-method}
\title{annotation_matches}
\usage{
annotation_matches(object)

\S4method{annotation_matches}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment with matches slot}
}
\value{
logical matrix of annotation matches
}
\description{
annotation_matches
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: get matches assay from SummarizedExperiment
}}

\author{
Alicia Schep
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)
matches <- annotation_matches(motif_ix)
}
