% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{annotationMatches}
\alias{annotationMatches}
\alias{annotationMatches,SummarizedExperiment-method}
\alias{annoation_matches<-,SummarizedExperiment-method}
\alias{annotationMatches<-}
\alias{annotationMatches<-,SummarizedExperiment-method}
\title{annotationMatches}
\usage{
annotationMatches(object)

annotationMatches(object) <- value

\S4method{annotationMatches}{SummarizedExperiment}(object)

\S4method{annotationMatches}{SummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment with matches slot, see details}

\item{value}{logical Matrix with annotation matches}
}
\value{
logical matrix of annotation matches
}
\description{
annotationMatches
}
\details{
Will extract matrix from the "matches", "annotationMatches", or
"motif_matches" assay of a SummarizedExperiment
}
\examples{
# load annotation matrix; result from matchMotifs
data(mini_ix, package = "chromVAR")
matches <- annotationMatches(mini_ix)
}
\author{
Alicia Schep
}
