% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{assemble_kmers}
\alias{assemble_kmers}
\title{assemble_kmers}
\usage{
assemble_kmers(object, threshold = 1.5, p = 0.01)
}
\arguments{
\item{object}{kmer deviations object}

\item{threshold}{variability threshold}

\item{p}{p value threshold for inclusion of kmer}
}
\value{
list with (1) motifs: de novo motif matrices, (2) seed: seed kmer for de novo motif
}
\description{
assemble_kmers
}
\details{
function for assembling de novo kmers from kmer deviations
}

