% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backgroundPeaks.R
\docType{class}
\name{backgroundPeaks-class}
\alias{backgroundPeaks}
\alias{backgroundPeaks-class}
\title{An S4 class that holds background peak set}
\description{
An S4 class that holds background peak set
}
\section{Slots}{

\describe{
\item{\code{peaks}}{}

\item{\code{background_peaks}}{}
}}

