% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filtering.R
\name{bias_filtering}
\alias{bias_filtering}
\title{bias_filtering}
\usage{
bias_filtering(counts, bias, norm = TRUE, what = c("both", "count", "bias"))
}
\arguments{
\item{counts}{counts}

\item{bias}{either vector with bias values for peaks, or GenomicRanges object with column named "gc"}

\item{norm}{count normalization}

\item{what}{filter based on gc bias, enrichment bias, or both? default is both}
}
\value{
vector of indices of peaks that pass bias filter
}
\description{
function to identify samples that show strong biases
}

