% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{cbind,chromVARDeviations-method}
\alias{cbind,chromVARDeviations-method}
\title{cbind method for chromVARDeviations}
\usage{
\S4method{cbind}{chromVARDeviations}(..., deparse.level = 1)
}
\arguments{
\item{...}{chromVARDeviations object to be combined}

\item{deparse.level}{See ?base::rbind for a description of this argument.}
}
\value{
chromVARDeviations object
}
\description{
cbind returns an error when applied to chromVARDeviations because results for
all cells or samples should originate from same computeDeviations computation
}
\seealso{
\code{\link{chromVARDeviations-class}}
}
\author{
Alicia Schep
}
