% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{chromVAR_deprecated}
\alias{chromVAR_deprecated}
\alias{assemble_kmers}
\alias{get_annotation_synergy}
\alias{get_annotation_correlation}
\alias{make_bias_bins}
\alias{make_permuted_sets}
\alias{pwm_distance}
\alias{plot_variability}
\alias{plot_deviations_tsne}
\alias{get_jaspar_motifs}
\alias{match_kmers}
\alias{deviations_covariability}
\alias{plot_kmer_mismatch}
\alias{get_total_fragments}
\alias{get_fragments_per_peak}
\alias{get_fragments_per_sample}
\alias{get_peaks}
\alias{get_counts}
\alias{get_sample_depths}
\alias{read_macs2_narrowpeaks}
\alias{get_annotations}
\alias{filter_samples}
\alias{filter_samples_plot}
\alias{filter_peaks}
\alias{annotation_matches}
\alias{deviations_tsne}
\alias{differential_deviations}
\alias{differential_variability}
\alias{compute_variability}
\alias{compute_expectations}
\alias{compute_deviations}
\alias{deviation_scores}
\alias{get_sample_distance}
\alias{get_sample_correlation}
\alias{get_cis_groups}
\alias{add_gc_bias}
\alias{get_background_peaks}
\alias{get_permuted_data}
\title{Deprecated functions in chromVAR}
\usage{
assemble_kmers(...)

get_annotation_synergy(...)

get_annotation_correlation(...)

make_bias_bins(...)

make_permuted_sets(...)

pwm_distance(...)

plot_variability(...)

plot_deviations_tsne(...)

get_jaspar_motifs(...)

match_kmers(...)

deviations_covariability(...)

plot_kmer_mismatch(...)

get_total_fragments(...)

get_fragments_per_peak(...)

get_fragments_per_sample(...)

get_peaks(...)

get_counts(...)

get_sample_depths(...)

read_macs2_narrowpeaks(...)

get_annotations(...)

filter_samples(...)

filter_samples_plot(...)

filter_peaks(...)

annotation_matches(...)

deviations_tsne(...)

differential_deviations(...)

differential_variability(...)

compute_variability(...)

compute_expectations(...)

compute_deviations(...)

deviation_scores(...)

get_sample_distance(...)

get_sample_correlation(...)

get_cis_groups(...)

add_gc_bias(...)

get_background_peaks(...)

get_permuted_data(...)
}
\arguments{
\item{...}{arguments passed to new function}
}
\value{
calls the replacement functions
}
\description{
chromVAR has moved functions and methods to camelCase from snake_case.
The following functions have been deprecated and replaced with a different
name:
\itemize{
  \item get_annotation_synergy is now \code{\link{getAnnotationSynergy}}
  \item make_bias_bins is now \code{\link{makeBiasBins}}
  \item make_permuted_sets is now \code{\link{makePermutedSets}}
  \item pwm_distance is now \code{\link{pwmDistance}}
  \item plot_variability is now \code{\link{plotVariability}} 
  \item plot_deviations_tsne is now \code{\link{plotDeviationsTsne}}
  \item get_jaspar_motifs is now \code{\link{getJasparMotifs}}
  \item match_kmers is now \code{\link{matchKmers}}
  \item deviations_covariability is now \code{\link{deviationsCovariability}}
  \item plot_kmer_mismatch is now \code{\link{plotKmerMismatch}}
  \item get_total_fragments is now \code{\link{getTotalFragments}}
  \item get_fragments_per_peak is now \code{\link{getFragmentsPerPeak}}
  \item get_fragments_per_sample is now \code{\link{getFragmentsPerSample}}
  \item get_peaks is now \code{\link{getPeaks}}
  \item get_counts is now \code{\link{getCounts}}
  \item get_sample_depths is now \code{\link{getSampleDepths}}
  \item read_macs2_narrowpeaks is now \code{\link{readNarrowpeaks}}
  \item get_annotations is now \code{\link{getAnnotations}}
  \item filter_samples is now \code{\link{filterSamples}}
  \item filter_samples_plot is now \code{\link{filterSamplesPlot}}
  \item filter_peaks is now \code{\link{filterPeaks}}
  \item annotation_matches is now \code{\link{annotationMatches}}
  \item deviations_tsne is now \code{\link{deviationsTsne}}
  \item differential_deviations is now \code{\link{differentialDeviations}}
  \item differential_variability is now \code{\link{differentialVariability}}
  \item compute_variability is now \code{\link{computeVariability}}
  \item compute_expectations is now \code{\link{computeExpectations}}
  \item computeDeviations is now \code{\link{computeDeviations}}
  \item deviation_scores is now \code{\link{deviationScores}}
  \item get_sample_distance is now \code{\link{getSampleDistance}}
  \item get_sample_correlation is now \code{\link{getSampleCorrelation}}
  \item get_cis_groups is now \code{\link{getCisGroups}}
  \item add_gc_bias is now \code{\link{addGCBias}}
  \item get_background_peaks is now \code{\link{getBackgroundPeaks}}
  \item get_permuted_data is now \code{\link{getPermutedData}}
  \item assemble_kmers is now \code{\link{assembleKmers}}
}
}
\author{
Alicia Schep
}
