% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{chromVAR_theme}
\alias{chromVAR_theme}
\title{chromVAR_theme}
\usage{
chromVAR_theme(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
ggplot2 theme
}
\description{
chromVAR_theme
}

