% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/compute_deviations.R
\docType{class}
\name{chromVARDeviations-class}
\alias{chromVARDeviations-class}
\alias{deviation_scores,chromVARDeviations-method}
\alias{deviations,chromVARDeviations-method}
\title{chromVARDeviations}
\usage{
\S4method{deviations}{chromVARDeviations}(object)

\S4method{deviation_scores}{chromVARDeviations}(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\description{
Class for storing results from \code{\link{compute_deviations}} function.
}
\details{
This class inherits from \code{\link[SummarizedExperiment]{SummarizedExperiment}}, and most methods for 
that class should work for objects of this class as well. Additionally, two
accessor functions are defined for extracting bias corrected deviations 
(\code{\link{deviations}}) and deviation Z-scores 
(\code{\link{deviation_scores}})
}
\section{Methods (by generic)}{
\itemize{
\item \code{deviations}: access deviation Z-scores matrix

\item \code{deviation_scores}: access deviation Z-scores matrix
}}

