% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResultSet.R
\name{cluster_samples}
\alias{cluster_samples}
\title{cluster_samples}
\usage{
cluster_samples(object, k = 5, plot = TRUE)
}
\arguments{
\item{object}{\code{\link{deviationResultSet}} object}

\item{k}{number of clusters}
}
\value{
list with two elements.  "hclust" is \code{\link[stats]{hclust}} object
"cluster" is cluster membership vector.
}
\description{
function to custer samples
}

