% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResultSet.R
\name{cluster_sets}
\alias{cluster_sets}
\title{cluster_sets}
\usage{
cluster_sets(object, k = NULL, max_k = 15, plot = TRUE)
}
\arguments{
\item{object}{\code{\link{deviationResultSet}} object}

\item{k}{number of clusters -- if not provided, will be estimated. See Details}

\item{max_k}{if k is not provided, max_k allowed}
}
\value{
list with two elements.  "hclust" is \code{\link[stats]{hclust}} object
"cluster" is cluster membership vector.
}
\description{
function to custer the top annotation sets
}

