% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{collapse_samples}
\alias{collapse_samples}
\title{collapse_samples}
\usage{
collapse_samples(counts_mat, grouping)
}
\arguments{
\item{counts_mat}{fragmentCounts object}

\item{grouping}{vector with groupings}
}
\value{
\code{\link{fragmentCounts}} object
}
\description{
collapse samples based on provided groupings
}
\details{
Meta data will be stripped from new fragmentCounts object
}
\seealso{
\code{\link{getFragmentCounts}}, \code{\link{cluster_samples}}, \code{\link{fragmentCounts}}
}

