% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bias.R
\docType{methods}
\name{compute_bias}
\alias{compute_bias}
\alias{compute_bias,GenomicRanges-method}
\alias{compute_bias,fragmentCounts-method}
\title{compute_bias}
\usage{
compute_bias(object, ...)

\S4method{compute_bias}{GenomicRanges}(object,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)

\S4method{compute_bias}{fragmentCounts}(object,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{object}{GenomicRanges or fragmentCounts}

\item{genome}{BSgenome object, e.g. BSgenome.Hsapiens.UCSC.hg19}
}
\value{
If object is GenomicRanges, returns GenomicRanges with additional meta-data
 column named "gc" with gc content

If object is fragmentCounts, returns fragmentCounts with additional meta-data
 column for peaks slot named "gc" with gc content
}
\description{
Computes GC content for peaks
}

