% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_deviations}
\alias{compute_deviations}
\title{compute_deviations}
\usage{
compute_deviations(peak_set, counts_mat, bg_peaks, niterations = 50,
  metric = c("z-score", "old"))
}
\arguments{
\item{peak_set}{}

\item{counts_mat}{}

\item{bg_peaks}{}

\item{niterations}{}

\item{metric}{}
}
\value{
\code{\link{deviationResult}}
}
\description{
Computes deviations across samples for a set of annotations
}
\seealso{
\code{\link{deviationResult}}, \code{\link{compute_variability}}
}

