% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_variability}
\alias{compute_variability}
\title{compute_variability}
\usage{
compute_variability(motif_indices, counts_mat, bg_peaks, niterations = 50,
  metric = c("z-score", "old"), BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{motif_indices}{}

\item{counts_mat}{}

\item{bg_peaks}{}

\item{niterations}{}

\item{metric}{}

\item{BPPARAM}{}
}
\value{
\code{\link{deviationResultSet}}
}
\description{
Computes variability across sets of annotations
}
\seealso{
\code{\link{deviationResultSet}}, \code{\link{variability}}, \code{\link{plot_variability}}
}

