% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs.R
\docType{methods}
\name{counts,SummarizedExperiment-method}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-,SummarizedExperiment,MatrixOrmatrix-method}
\alias{counts<-,SummarizedExperiment-method}
\title{Accessors for the 'counts' slot of a SummarizedExperiment}
\usage{
\S4method{counts}{SummarizedExperiment}(object)

\S4method{counts}{SummarizedExperiment,MatrixOrmatrix}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment object}

\item{value}{matrix of counts}
}
\value{
Matrix of counts
}
\description{
Accessors for the 'counts' slot of a SummarizedExperiment
}
\examples{
data(mini_counts, package = "chromVAR")
fragment_counts <- counts(mini_counts)
}
