% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{deviation_scores}
\alias{deviation_scores}
\alias{deviation_scores,chromVARDeviations-method}
\alias{deviation_scores,chromVARDeviations-method}
\title{deviation_scores}
\usage{
deviation_scores(object)

\S4method{deviation_scores}{chromVARDeviations}(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\value{
The deviation_scores and deviations accessors both return matrices.

matrix of deviation Z-scores
}
\description{
Accessor for deviation Z-scores from \code{\link{chromVARDeviations-class}} object
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo
library(motifmatchr)
library(BSgenome.Hsapiens.UCSC.hg19)
motif_ix <- match_motifs(motifs, mini_counts,
                         genome = BSgenome.Hsapiens.UCSC.hg19)

# computing deviations
dev <- compute_deviations(object = mini_counts,
                         annotations = motif_ix)
scores <- deviation_scores(dev)
}
\author{
Alicia Schep
}
