% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legacy.R
\docType{class}
\name{deviationBackgroundParameters-class}
\alias{deviationBackgroundParameters}
\alias{deviationBackgroundParameters-class}
\title{deviationBackgroundParameters}
\description{
A legacy class designed to facilitate computation of sampled background counts for a
set of annotations as in Buenrostro et al. (2015).
}
\section{Slots}{

\describe{
\item{\code{counts_sort}}{}

\item{\code{counts_sort_rev}}{}

\item{\code{bias_sort}}{}

\item{\code{bias_sort_rev}}{}

\item{\code{window}}{}

\item{\code{npeaks}}{}
}}

