% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R
\docType{class}
\name{deviationResult-class}
\alias{deviationResult}
\alias{deviationResult-class}
\alias{deviations,}
\alias{metric,}
\alias{name,}
\alias{pvalues}
\alias{variability,}
\alias{variability_bounds,}
\title{An S4 class to store chromVAAR result}
\description{
An S4 class to store chromVAAR result
}
\section{Slots}{

\describe{
\item{\code{deviations}}{}

\item{\code{variability}}{}

\item{\code{variability_bounds}}{}

\item{\code{name}}{}

\item{\code{metric}}{}

\item{\code{pvalues}}{}
}}

