% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{deviations}
\alias{deviations}
\alias{deviations,chromVARDeviations-method}
\title{deviations}
\usage{
deviations(object)

\S4method{deviations}{chromVARDeviations}(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\value{
matrix of bias corrected deviations
}
\description{
Accessor for bias corrected deviations from 
\code{\link{chromVARDeviations-class}} object
}
\examples{
# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")
bias_corrected_deviations <- deviations(mini_dev)
}
\author{
Alicia Schep
}
