% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{deviations_covariability}
\alias{deviations_covariability}
\title{deviations_covariability}
\usage{
deviations_covariability(object)
}
\arguments{
\item{object}{deviations result}
}
\value{
'covariability' matrix
}
\description{
deviations_covariability
}
\details{
Returns the 'covariability' between motifs/kmers/peaksets. 
Covariability' is defined as covariance between Z-scores divided by variance 
of Z-scores for one motif/kmer/peakset (the row).
}
\examples{
# load very small example data
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
                         
# get covariability for just first three motifs                         
devcov <- deviations_covariability(dev[1:3,])                         
}
