% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{deviations_tsne_motifs}
\alias{deviations_tsne_motifs}
\title{deviations_tsne_motifs}
\usage{
deviations_tsne_motifs(motifs = NULL, kmers = NULL, threshold = 1.5,
  perplexity = 10, shiny = interactive())
}
\arguments{
\item{threshold}{variability threshold -- use only deviations with variability greater than threshold}

\item{perplexity}{perplexity parameter for tsne}

\item{shiny}{output shiny gadget for choosing parameters & visualizing results}

\item{object}{deviations result}
}
\value{
list with three elements: threshold used, perplexity used, tsne results
from \code{\link[Rtsne]{Rtsne}}
}
\description{
deviations_tsne_motifs
}

