% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{differential_deviations}
\alias{differential_deviations}
\title{differential_deviations}
\usage{
differential_deviations(object, groups, alternative = c("two.sided", "less",
  "greater"), parametric = TRUE)
}
\arguments{
\item{object}{chromVARDeviations object}

\item{groups}{either vector of groups or name of column in colData of object with grouop information}

\item{alternative}{only used if there are two groups -- two.sided or one sided test}

\item{parametric}{use parametric test. alternatively will use kruskal wallace}
}
\value{
data.frame with p value and adjusted p value
}
\description{
Function to see whether deviations differ between groups
}

