% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R
\name{differential_peaks}
\alias{differential_peaks}
\title{differential_peaks}
\usage{
differential_peaks(object, ...)
}
\arguments{
\item{object}{\code{\link{deviationResult}} object}
}
\description{
get differential_peaks
}
\seealso{
\code{\link{get_pvalues}}, \code{\link{metric}},
\code{\link{variability}}, \code{\link{deviations}}, \code{\link{differential_samples}}
}

