% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{differential_variability}
\alias{differential_variability}
\title{differential_variability}
\usage{
differential_variability(object, groups, parametric = TRUE)
}
\arguments{
\item{object}{chromVARDeviations object}

\item{groups}{either vector of groups or name of column in colData of object with grouop information}

\item{parametric}{use parametric test. alternatively will use kruskal wallace}
}
\value{
data.frame with p value and adjusted p value
}
\description{
Function to determine whether groups differ in variability
}

