% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{differentialVariability}
\alias{differentialVariability}
\title{differentialVariability}
\usage{
differentialVariability(object, groups, parametric = TRUE)
}
\arguments{
\item{object}{chromVARDeviations object}

\item{groups}{either vector of groups or name of column in colData of object 
with grouop information}

\item{parametric}{use parametric test. alternatively will use kruskal wallace}
}
\value{
data.frame with p value and adjusted p value
}
\description{
Function to determine whether groups differ in variability
}
\examples{
# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")
difvar <- differentialVariability(mini_dev, "Cell_Type")
}
\author{
Alicia Schep
}
