% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{example_counts}
\format{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment}}}
\usage{
data(example_counts)
}
\description{
Very small sample data set for trying out chromVAR
}
\keyword{datasets}

