% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{filter_samples}
\alias{filter_samples}
\title{filter_samples}
\usage{
filter_samples(counts_mat, depths, min_in_peaks = 0.25,
  min_fragments = 1000)
}
\arguments{
\item{counts_mat}{matrix of fragment counts per peak per sample, as computed
by \code{\link{getFragmentCounts}}}

\item{depths}{vector of sequencing depth per samples, as computed by
\code{\link{getSampleDepths}}}

\item{min_in_peaks}{minimum fraction of samples within peaks}

\item{min_fragments}{minimum number of fragments within peaks}
}
\value{
vector of indices, representing samples that should be kept
}
\description{
function to get indices of samples that pass filtters
}
\seealso{
\code{\link{get_counts}},  \code{\link{get_inputs}}, \code{\link{filter_peaks}}
}

