% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_samples_plot}
\alias{filter_samples_plot}
\title{filter_samples_plot}
\usage{
filter_samples_plot(object, min_in_peaks = NULL, min_depth = NULL,
  use_plotly = interactive())
}
\arguments{
\item{object}{SummarizedExperiment with matrix of fragment counts per peak 
per sample, as computed by \code{\link{get_counts}}}

\item{min_in_peaks}{minimum fraction of samples within peaks}

\item{min_depth}{minimum library size}

\item{use_plotly}{make interactive plot?}
}
\value{
indices of samples to keep
}
\description{
plot filtering of samples
}
\details{
If unspecified, min_in_peaks and min_depth cutoffs will be estimated
based on data. min_in_peaks is set to 0.5 times the median proportion of 
fragments in peaks.  min_depth is set to the maximum of 500 or 10% of the 
median library size.
}
\seealso{
\code{\link{get_counts}},  \code{\link{get_peaks}}, 
\code{\link{filter_peaks}}
}

