% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{filterFragmentCounts}
\alias{filterFragmentCounts}
\title{filterFragmentCounts}
\usage{
filterFragmentCounts(counts_mat, min_in_peaks = 0.25, min_fragments = 2500,
  min_fragments_per_peak = 3)
}
\arguments{
\item{counts_mat}{fragmentCounts object}

\item{min_in_peaks}{minimum fraction of fragments within peaks for a sample (default = 0.25)}

\item{min_fragments}{minimum fragments in peaks for a sample(default = 2500)}

\item{min_fragments_per_peak}{minimum fragments across samples that fall into a peak}
}
\value{
\code{\link{fragmentCounts}} object
}
\description{
Filter Fragment Counts by quality
}
\details{
Filtering of samples based on min_in_peaks and min_fragments arguments is performed first.
Then peaks are filtered by min_fragments_per_peak argument (so only fragments in samples that survived
filtering of samples are counted).
}
\seealso{
\code{\link{getFragmentCounts}}, \code{\link{getFragmentCountsByRG}}, \code{\link{fragmentCounts}}
}

