% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{foldchange}
\alias{foldchange}
\alias{foldchange,deviationResult-method}
\alias{foldchange,deviationResultSet-method}
\title{foldchange}
\usage{
foldchange(object, ...)

\S4method{foldchange}{deviationResult}(object)

\S4method{foldchange}{deviationResultSet}(object)
}
\arguments{
\item{object}{either \code{\link{deviationResult}} or \code{\link{deviationResultSet}}}
}
\description{
Accessor for foldchange in deviationResult.
}
\details{
foldchange returns the fold change difference between the observed accessibility
for a set of peaks in a particular sample and the mean accessibility in sets of
background peaks
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns vector of fold change for samples

\item \code{deviationResultSet}: returns matrix of fold changes, with columns representing
samples and rows representing annotation sets
}}
\seealso{
\code{\link{deviations}}, \code{\link{variability_bounds}}
}

