% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\docType{class}
\name{fragmentCounts-class}
\alias{[,fragmentCounts,ANY,ANY,ANY-method}
\alias{fragmentCounts}
\alias{fragmentCounts-class}
\alias{initialize,fragmentCounts-method}
\alias{show,fragmentCounts-method}
\title{fragmentCounts}
\usage{
\S4method{initialize}{fragmentCounts}(.Object, ...)

\S4method{show}{fragmentCounts}(object)

\S4method{[}{fragmentCounts,ANY,ANY,ANY}(x, i, j)
}
\description{
fragmentCounts is a class to store fragment counts for genomic regions across samples.
The class is specifically designed to be able to handle sparse data, as would
arise from single-cell ATAC-seq experiment.  The counts slot holds the count data as
a Matrix (from package Matrix); the specific class of the Matrix is determined by
sparsity of the data.  The fragmentCounts class also stores the genomic regions that
correspond to the rows of the counts matrix. Optionally, the class can store sample meta data
as a data.frame with rows corresponding to samples.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Matrix of fragment counts, with each row representing a peak, each column a sample}

\item{\code{peaks}}{GRanges object with peaks for each row of counts}

\item{\code{total_fragments}}{total fragments within peaks}

\item{\code{fragments_per_sample}}{total fragments within peaks per sample}

\item{\code{fragments_per_peak}}{total fragments within each peak across samples}

\item{\code{nsample}}{number of samples}

\item{\code{npeak}}{number of peaks}

\item{\code{sample_meta}}{optional data.frame with meta data for samples, with rows corresponding to samples}

\item{\code{depth}}{total read depth for each sample}
}}
\section{Accessors}{

Accessors are available for counts, peaks, and meta.
}

