% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{fragmentCounts}
\alias{fragmentCounts}
\title{An S4 class that holds fragment counts across peak regions}
\usage{
fragmentCounts(counts, peaks, depth)

fragmentCounts(counts, peaks, depth)
}
\arguments{
\item{counts}{A Matrix with fragment counts-- rows should correspond to peaks, columns to samples}

\item{peaks}{GRanges with peaks}
}
\description{
An S4 class that holds fragment counts across peak regions

makes fragmentCounts object using counts Matrix and peaks.
generally not for use by end-user.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Matrix of fragment counts, with each row representing a peak, each column a sample}

\item{\code{peaks}}{GRanges object with peaks for each row of counts}

\item{\code{total_fragments}}{total fragments within peaks}

\item{\code{fragments_per_sample}}{total fragments within peaks per sample}

\item{\code{fragments_per_peak}}{total fragments within each peak across samples}

\item{\code{nsample}}{number of samples}

\item{\code{npeak}}{number of peaks}

\item{\code{sample_meta}}{optional data.frame with meta data for samples, with rows corresponding to samples}

\item{\code{depth}}{total read depth for each sample}
}}

