% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_synergy.R
\docType{methods}
\name{get_annotation_synergy}
\alias{get_annotation_synergy}
\alias{get_annotation_synergy,SummarizedExperiment,SummarizedExperiment-method}
\alias{get_annotation_synergy,SummarizedExperiment,MatrixOrmatrix-method}
\alias{get_annotation_synergy,SummarizedExperiment,list-method}
\alias{get_annotation_synergy,MatrixOrmatrix,SummarizedExperiment-method}
\alias{get_annotation_synergy,MatrixOrmatrix,MatrixOrmatrix-method}
\alias{get_annotation_synergy,MatrixOrmatrix,list-method}
\title{get_annotation_synergy}
\usage{
get_annotation_synergy(object, annotations, ...)


  \S4method{get_annotation_synergy}{SummarizedExperiment,SummarizedExperiment}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object), variabilities = NULL,
  nbg = 25)

\S4method{get_annotation_synergy}{SummarizedExperiment,MatrixOrmatrix}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object), variabilities = NULL,
  nbg = 25)

\S4method{get_annotation_synergy}{SummarizedExperiment,list}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object), variabilities = NULL,
  nbg = 25)

\S4method{get_annotation_synergy}{MatrixOrmatrix,SummarizedExperiment}(object,
  annotations, background_peaks, expectation = compute_expectations(object),
  variabilities = NULL, nbg = 25)

\S4method{get_annotation_synergy}{MatrixOrmatrix,MatrixOrmatrix}(object,
  annotations, background_peaks, expectation = compute_expectations(object),
  variabilities = NULL, nbg = 25)

\S4method{get_annotation_synergy}{MatrixOrmatrix,list}(object, annotations,
  background_peaks, expectation = compute_expectations(object),
  variabilities = NULL, nbg = 25)
}
\arguments{
\item{object}{result from compute_deviations}

\item{annotations}{SummarizedExperiment of annotation matches}

\item{...}{additional arguments}

\item{background_peaks}{optional, matrix of background peaks}

\item{expectation}{optional, expected fraction of reads per peak, as computed
by compute_expectations}

\item{variabilities}{optional, variabilities computed from compute_variability}

\item{nbg}{number of background iterations}
}
\value{
synergy matrix
}
\description{
get_annotation_synergy
}
\details{
should only be run on small number of motifs/kmers/peaksets (very slow!)
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SummarizedExperiment,annotations = SummarizedExperiment}: object and annotations are SummarizedExperiment

\item \code{object = SummarizedExperiment,annotations = MatrixOrmatrix}: object is SummarizedExperiment, 
annotations are Matrix

\item \code{object = SummarizedExperiment,annotations = list}: object is SummarizedExperiment, 
annotations are list

\item \code{object = MatrixOrmatrix,annotations = SummarizedExperiment}: object and annotations are SummarizedExperiment

\item \code{object = MatrixOrmatrix,annotations = MatrixOrmatrix}: object is SummarizedExperiment, 
annotations are Matrix

\item \code{object = MatrixOrmatrix,annotations = list}: object is SummarizedExperiment, 
annotations are list
}}

