% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\docType{methods}
\name{get_annotations}
\alias{get_annotations}
\alias{get_annotations,GRangesList-method}
\alias{get_annotations,MatrixOrmatrix-method}
\alias{get_annotations,data.frame-method}
\alias{get_annotations,list-method}
\alias{get_annotations,character-method}
\title{get_annotations}
\usage{
get_annotations(annotations, ...)

\S4method{get_annotations}{GRangesList}(annotations, rowRanges, ...)

\S4method{get_annotations}{MatrixOrmatrix}(annotations, ...)

\S4method{get_annotations}{data.frame}(annotations, ...)

\S4method{get_annotations}{list}(annotations, npeaks = NULL, ...)

\S4method{get_annotations}{character}(annotations, rowRanges, column = NULL,
  colData = NULL)
}
\arguments{
\item{annotations}{matrix, Matrix, or data.frame of fragment counts,
or SummarizedExperiment with counts assays, see details}

\item{...}{additional arguments}

\item{rowRanges}{GenomicRanges or GenomicRangesList or 
RangedSummarizedExperiment}

\item{npeaks}{number of peaks}

\item{column}{column of bed file with annotation names, see details}

\item{colData}{data on each annotation to store along with annotation matrix}
}
\value{
SummarizedExperiment object with 'matches' assay
}
\description{
get_annotations
}
\section{Methods (by class)}{
\itemize{
\item \code{GRangesList}: get annotation matrix from GRangesList

\item \code{MatrixOrmatrix}: get annotation matrix from Matrix or matrix

\item \code{data.frame}: get annotation matrix from data.frame

\item \code{list}: get annotation matrix from list

\item \code{character}: get annotations from bed files
}}

\examples{

# First get example counts
data(mini_counts, package = "chromVAR")

# Get annotations from genomic ranges list
library(GenomicRanges)
library(SummarizedExperiment)
my_annotation_granges <- GRangesList(GRanges("chr1", 
                                             ranges = IRanges(start = 
                                             c(566763,805090), width = 8)),
                                     GRanges("chr1", ranges = IRanges(start = 
                                               c(566792,895798), width = 8)))
anno_ix <- get_annotations(my_annotation_granges, 
                           rowRanges = rowRanges(mini_counts))
}
\author{
Alicia Schep
}
