% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/background_peaks.R
\name{get_background_peaks}
\alias{get_background_peaks}
\title{get_background_peaks}
\usage{
get_background_peaks(counts_mat, bias, niterations = 50, window = 500,
  with_replacement = TRUE)
}
\arguments{
\item{counts_mat}{counts matrix}

\item{bias}{either vector with bias values for peaks, or GenomicRanges object with column named "gc"}

\item{niterations}{number of background peaks to sample}

\item{window}{window size around peak in which to sample a background peak}

\item{with_replacement}{sample peaks with replacement? Default is TRUE}
}
\value{
matrix with one row per peak and one column per iteration.  values in a row
represent indices of background peaks for the peak with that index
}
\description{
Function to get a set of background peaks for each peak based on GC content and # of counts
across all samples
}
\details{
Background peaks are chosen by finding the window nearest neighbors to a peak in terms of
GC content and # of fragments across samples using the Mahalanobis distance.  From those nearest
neighbors, niterations peaks are sampled from that background.
}

