% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\docType{methods}
\name{get_cis_groups}
\alias{get_cis_groups}
\alias{get_cis_groups,GenomicRanges-method}
\alias{get_cis_groups,RangedSummarizedExperiment-method}
\title{get_cis_groups}
\usage{
get_cis_groups(object, ...)

\S4method{get_cis_groups}{RangedSummarizedExperiment}(object, grpsize = 25,
  stepsize = 10)

\S4method{get_cis_groups}{GenomicRanges}(object, grpsize = 25,
  stepsize = 10)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment}

\item{...}{additional arguments}

\item{grpsize}{number of peaks to include in each grouop}

\item{stepsize}{number of peaks between each new set of groups}
}
\description{
get_cis_groups
}
\section{Methods (by class)}{
\itemize{
\item \code{RangedSummarizedExperiment}: method for RangedSummarizedExperiment

\item \code{GenomicRanges}: method for GenomicRanges
}}

