% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{get_counts}
\alias{get_counts}
\title{getFragmentCounts}
\usage{
get_counts(bams, peaks, paired, by_rg = FALSE)
}
\arguments{
\item{bams}{filenames for bam file with aligned reads}

\item{peaks}{GRanges object with peaks}

\item{paired}{paired end data?}

\item{by_rg}{use RG tags to separate groups?}
}
\value{
\code{\link[Matrix]{Matrix}} object
}
\description{
makes matrix of fragment counts in peaks using one or multiple bam files
}
\seealso{
\code{\link{get_sample_depths}},  \code{\link{get_inputs}}, \code{\link{filter_samples}}
}

