% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/background_peaks.R
\name{get_gc}
\alias{get_gc}
\title{get_gc}
\usage{
get_gc(peaks,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{peaks}{GenomicRanges}

\item{genome}{BSgenome object}
}
\description{
Computes GC content for peaks
}

