% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kmer.R
\name{get_kmer_indices}
\alias{get_kmer_indices}
\title{get_kmer_indices}
\usage{
get_kmer_indices(peaks,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19, k = 6)
}
\arguments{
\item{genome}{\code{\link[BSgenome]{BSgenome-class}} object, default is hg19}

\item{k}{length of kmer, default is 6, must be between 5 and 8}

\item{peak}{\code{\link[GenomicRanges]{GenomicRanges}} object}
}
\value{
list of vectors with indices of peaks containing kmer
}
\description{
For each possible kmer of size k, finds which peaks contain 1 or more copies
of that kmer
}

