% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifs.R
\name{get_max_motif_scores}
\alias{get_max_motif_scores}
\title{get_max_motif_scores}
\usage{
get_max_motif_scores(motif, peaks,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{motif}{\code{\link[TFBSTools]{PFMatrix}} or \code{\link[TFBSTools]{PWMatrix}}}

\item{peaks}{\code{\link[GenomicRanges]{GenomicRanges}}}

\item{genome}{\code{\link[BSgenome]{BSgenome}} object}
}
\value{
A numeric vector with the max motif score for each peak
}
\description{
Function to get max motif score within each peak
}

