% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_peaks.R
\name{get_permuted_data}
\alias{get_permuted_data}
\title{get_permuted_data}
\usage{
get_permuted_data(object, niterations = 10, w = 0.1, bs = 50)
}
\arguments{
\item{niterations}{number of background peaks to sample}

\item{w}{parameter controlling similarity of background peaks}

\item{bs}{bin size parameter}

\item{counts_mat}{SummarizedExperiment}
}
\value{
matrix with one row per peak and one column per iteration.  values in a row
represent indices of background peaks for the peak with that index
}
\description{
Function to get permuted data while maintaining biases
}
\details{
Background peaks are chosen by sampling peaks based on similarity in
GC content and # of fragments across samples using the Mahalanobis distance.
The w paramter controls how similar background peaks should be.  The bs parameter
controls the precision with which the similarity is computed; increasing bs will
make the function run slower. Sensible default parameters are chosen for both.
}

