% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{get_pvalues}
\alias{get_pvalues}
\alias{get_pvalues,deviationResult-method}
\alias{get_pvalues,deviationResultSet-method}
\title{get_pvalues}
\usage{
get_pvalues(object, ...)

\S4method{get_pvalues}{deviationResult}(object, what = c("deviations",
  "variability"), adjust = TRUE)

\S4method{get_pvalues}{deviationResultSet}(object, adjust = TRUE)
}
\arguments{
\item{object}{either \code{\link{deviationResult}} or \code{\link{deviationResultSet}} object}

\item{what}{return p values for deviations or variability?}

\item{adjust}{adjust for multiple testing?  Will use Benjamini-Hochberg correction}
}
\description{
Accessor for p values
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns p values for either deviations or variability

\item \code{deviationResultSet}: returns p values for variability for each set
}}
\seealso{
\code{\link{metric}}, \code{\link{variability}}, \code{\link{deviations}}
}

