% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_sample_correlation}
\alias{get_sample_correlation}
\title{get_sample_correlation}
\usage{
get_sample_correlation(object, threshold = 1.5)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{threshold for variability}
}
\value{
correlation matrix between samples
}
\description{
Get correlation between samples based on bias corrected deviations
}
\details{
This function will compute the correlation between samples based on 
the normalized deviations. It will first remove correlated motifs/peak sets. 
Then the pearson correlation coefficient will be computed and returned.
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
sample_cor <- get_sample_correlation(dev, threshold = 1)  
# setting very low variabilitiy threshold because this is mini data set
# Use plot_variability to get a sense of an appropriate threshold
# As this is mini data set, results probably not very meaningful!
}
\seealso{
\code{\link{get_sample_distance}}
}
\author{
Alicia Schep
}
