% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_sample_distance}
\alias{get_sample_distance}
\title{get_sample_distance}
\usage{
get_sample_distance(object, threshold = 1.5, initial_dims = 50,
  distance_function = dist)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{threshold for variability}

\item{initial_dims}{initial dimentions for preliminary dimensionality 
reduction via pca}

\item{distance_function}{distance function to use}
}
\value{
dist object for distance between samples
}
\description{
Get distance between samples based on bias corrected deviations
}
\details{
This function will compute the distance between samples based on the 
normalized deviations.  It will first remove correlated motifs / peak sets. 
Then the dimensionality will be further reduced via PCA if the number of 
dimensions exceeds initial_dims.  Then  the supplied distance_function will 
be used.
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
sample_dist <- get_sample_distance(dev, threshold = 1)  
# setting very low variabilitiy threshold because this is mini data set
# Use plot_variability to get a sense of an appropriate threshold
# As this is mini data set, results probably not very meaningful!
}
\seealso{
\code{\link{get_sample_correlation}}
}
\author{
Alicia Schep
}
