% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{get_total_fragments}
\alias{get_total_fragments}
\alias{get_total_fragments,SummarizedExperiment-method}
\alias{get_total_fragments,MatrixOrmatrix-method}
\title{get_total_fragments}
\usage{
get_total_fragments(object)

\S4method{get_total_fragments}{SummarizedExperiment}(object)

\S4method{get_total_fragments}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
sum of all counts within object
}
\description{
get_total_fragments
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
total_frags <- get_total_fragments(mini_counts)
}
\seealso{
\code{\link{get_fragments_per_sample}}, 
\code{\link{get_fragments_per_peak}}
}
