% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variability_extensions.R
\name{get_variable_peaks}
\alias{get_variable_peaks}
\title{get_variable_peaks}
\usage{
get_variable_peaks(counts_mat, deviation_result)
}
\arguments{
\item{counts_mat}{\code{\link{fragmentCounts}} object}

\item{deviation_result}{\code{\link{deviationResult}} object}
}
\description{
function to find individual peaks with variability correlated with the variability
of a set of peaks
}

