% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{getFragmentCounts}
\alias{getFragmentCounts}
\title{make fragmentCounts object using multiple bam files}
\usage{
getFragmentCounts(bams, peaks)
}
\arguments{
\item{bams}{filenames for bam file with aligned reads}

\item{peaks}{GRanges object with peaks}
}
\description{
make fragmentCounts object using multiple bam files
}

