% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{getFragmentCountsByRG}
\alias{getFragmentCountsByRG}
\title{make fragmentCounts object using bam file with RG tags}
\usage{
getFragmentCountsByRG(bam, peaks, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{bam}{filename for bam file with aligned reads}

\item{peaks}{GRanges object with peaks}
}
\description{
make fragmentCounts object using bam file with RG tags
}

