% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{human_pwms_v1}
\alias{human_pwms_v1}
\title{human_pwms_v1}
\format{\code{\link[TFBSTools]{XMatrixList}} of length 1764}
\usage{
data(human_pwms_v1)
}
\description{
Collection of human pwms
}
\keyword{datasets}

