% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\docType{methods}
\name{make_bias_bins}
\alias{make_bias_bins}
\alias{make_bias_bins,MatrixOrmatrix-method}
\alias{make_bias_bins,RangedSummarizedExperiment-method}
\alias{make_bias_bins,SummarizedExperiment-method}
\title{make_bias_bins}
\usage{
make_bias_bins(object, ...)

\S4method{make_bias_bins}{SummarizedExperiment}(object,
  bias = rowData(object)$bias, nbins = 25, frac = 0.3)

\S4method{make_bias_bins}{RangedSummarizedExperiment}(object,
  bias = rowRanges(object)$bias, nbins = 25, frac = 0.3)

\S4method{make_bias_bins}{MatrixOrmatrix}(object, bias, nbins = 25,
  frac = 0.3)
}
\arguments{
\item{object}{fragment counts stored as RangedSummarizedExperiment, 
SummarizedExperiment, matrix, or Matrix}

\item{...}{addtional arguments}

\item{bias}{vector of some bias signal (usually gc content) for each row of 
object}

\item{nbins}{number of bins for each category, see Details}

\item{frac}{fraction of peaks within given bin to select randomly}
}
\value{
SummarizedExperiment storing bias bins annotation
}
\description{
Makes bins based on fragment counts
}
\details{
Will create nbins * 3 annotations based on sampling from peaks with 
a certain fragment count, fragment count, or fragment count & bias.
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment

\item \code{RangedSummarizedExperiment}: method for RangedSummarizedExperiment

\item \code{MatrixOrmatrix}: method for Matrix or matrix
}}

