% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sets.R
\docType{methods}
\name{make_permuted_sets}
\alias{make_permuted_sets}
\alias{make_permuted_sets,SummarizedExperiment,SummarizedExperiment-method}
\alias{make_permuted_sets,RangedSummarizedExperiment,SummarizedExperiment-method}
\alias{make_permuted_sets,MatrixOrmatrix,SummarizedExperiment-method}
\alias{make_permuted_sets,SummarizedExperiment,MatrixOrmatrix-method}
\alias{make_permuted_sets,RangedSummarizedExperiment,MatrixOrmatrix-method}
\alias{make_permuted_sets,MatrixOrmatrix,MatrixOrmatrix-method}
\alias{make_permuted_sets,SummarizedExperiment,list-method}
\alias{make_permuted_sets,RangedSummarizedExperiment,list-method}
\alias{make_permuted_sets,MatrixOrmatrix,list-method}
\title{make_permuted_sets}
\usage{
make_permuted_sets(object, annotations, ...)


  \S4method{make_permuted_sets}{SummarizedExperiment,SummarizedExperiment}(object,
  annotations, bias = rowData(object)$bias, window = 10)


  \S4method{make_permuted_sets}{RangedSummarizedExperiment,SummarizedExperiment}(object,
  annotations, bias = rowRanges(object)$bias, window = 10)

\S4method{make_permuted_sets}{MatrixOrmatrix,SummarizedExperiment}(object,
  annotations, bias, window = 10)

\S4method{make_permuted_sets}{SummarizedExperiment,MatrixOrmatrix}(object,
  annotations, bias = rowData(object)$bias, window = 10)


  \S4method{make_permuted_sets}{RangedSummarizedExperiment,MatrixOrmatrix}(object,
  annotations, bias = rowRanges(object)$bias, window = 10)

\S4method{make_permuted_sets}{MatrixOrmatrix,MatrixOrmatrix}(object,
  annotations, bias, window = 10)

\S4method{make_permuted_sets}{SummarizedExperiment,list}(object, annotations,
  bias = rowData(object)$bias, window = 10)

\S4method{make_permuted_sets}{RangedSummarizedExperiment,list}(object,
  annotations, bias = rowRanges(object)$bias, window = 10)

\S4method{make_permuted_sets}{MatrixOrmatrix,list}(object, annotations, bias,
  window = 10)
}
\arguments{
\item{object}{fragment counts stored as RangedSummarizedExperiment, 
SummarizedExperiment, matrix, or Matrix}

\item{annotations}{annotations as SummarizedExperiment, matrix, or list}

\item{...}{additional arguments}

\item{bias}{vector of some bias signal (usually gc content) for each row of 
object}

\item{window}{number of nearest neighbors to consider}
}
\value{
SummarizedExperiment storing bias bins annotation
}
\description{
Makes annotations sets with similar bias to input sets
}
\details{
Will create nbins * 3 annotations based on sampling from peaks with 
a certain fragment count, fragment count, or fragment count & bias.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SummarizedExperiment,annotations = SummarizedExperiment}: method for SummarizedExperiment and SummarizedExperiment

\item \code{object = RangedSummarizedExperiment,annotations = SummarizedExperiment}: method for RangedSummarizedExperiment and SummarizedExperiment

\item \code{object = MatrixOrmatrix,annotations = SummarizedExperiment}: method for Matrix or matrix and SummarizedExperiment

\item \code{object = SummarizedExperiment,annotations = MatrixOrmatrix}: method for SummarizedExperiment and MatrixOrmatrix

\item \code{object = RangedSummarizedExperiment,annotations = MatrixOrmatrix}: method for RangedSummarizedExperiment and MatrixOrmatrix

\item \code{object = MatrixOrmatrix,annotations = MatrixOrmatrix}: method for Matrix/matrix and Matrix/matrix

\item \code{object = SummarizedExperiment,annotations = list}: method for SummarizedExperiment and list

\item \code{object = RangedSummarizedExperiment,annotations = list}: method for RangedSummarizedExperiment and list

\item \code{object = MatrixOrmatrix,annotations = list}: method for Matrix or matrix and list
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

perm_sets <- make_permuted_sets(mini_counts, motif_ix)
}
\author{
Alicia Schep
}
