% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_kmers.R
\name{match_kmers}
\alias{match_kmers}
\title{match_kmers}
\usage{
match_kmers(k, subject, ...)
}
\arguments{
\item{k}{k}

\item{subject}{either \code{\link[GenomicRanges]{GenomicRanges}}, \code{\link[Biostrings]{DNAStringSet}},
\code{\link[Biostrings]{DNAString}}, or character vector}

\item{genome}{BSgenome object, only used if subect is \code{\link[GenomicRanges]{GenomicRanges}}}

\item{out}{what to return? see details}

\item{if}{subject is not GenomicRanges, ranges to use when out is positions}
}
\description{
Find kmer matches
}
\details{
Can either return a SummarizedExperiment with just sparse matrix with values set to 1 for a match (if return == "matches"),  or
a GenomicRanges  object with all the positions of matches
}

