% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{mini_counts}
\alias{mini_counts}
\title{mini_counts}
\usage{
data(mini_counts)
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment}}
}
\description{
Tiny sample data set for chromVAR funtion examples
}
\examples{
data(mini_counts)
}
\seealso{
\code{\link{mini_dev}}, \code{\link{mini_ix}}
}
\keyword{datasets}
