% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{mini_dev}
\alias{mini_dev}
\title{mini_dev}
\usage{
data(mini_dev)
}
\value{
\code{\link{chromVARDeviations-class}}
}
\description{
Tiny sample chromVARDeviations object resulting from computeDeviations
Result from running computeDeviations(mini_counts, mini_ix) on 
mini_ix and mini_counts data from this package
}
\examples{
data(mini_dev)
}
\seealso{
\code{\link{computeDeviations}}, \code{\link{mini_counts}},
\code{\link{mini_ix}}
}
\keyword{datasets}
