% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{mini_ix}
\alias{mini_ix}
\title{mini_ix}
\usage{
data(mini_ix)
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment}}
}
\description{
Tiny sample annotation object for use in chromVAR examples
Result from running matchMotifs(example_motifs,mini_counts,"hg19) on 
example_motifs from motifmatchr package and mini_counts from this package
}
\examples{
data(mini_ix)
}
\seealso{
\code{\link{mini_counts}}, \code{\link{mini_dev}}
}
\keyword{datasets}
