% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deviations}
\alias{plot_deviations}
\title{plot_deviations}
\usage{
plot_deviations(X, xlabel = NA, ylabel = NA, name = NA,
  cluster_row = TRUE, cluster_col = TRUE, cluster_row_dist = cor_dist,
  cluster_col_dist = cor_dist, sample_annotation = NULL, set_names = NULL)
}
\arguments{
\item{X}{deviations matrix (either z-score or fold change)}

\item{xlabel}{label for x axis}

\item{ylabel}{label for y axis}

\item{name}{title for plot}

\item{cluster_row}{cluster rows?}

\item{cluster_col}{cluster columns?}

\item{cluster_row_dist}{distance function to use for clustering rows}

\item{cluster_col_dist}{distance function to use for clustering columns}

\item{sample_annotation}{annotation of sample}
}
\value{
ggplot object
}
\description{
plot heatmap of deviations
}

