% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deviations_heatmap}
\alias{plot_deviations_heatmap}
\title{plot_deviations}
\usage{
plot_deviations_heatmap(object, top = 100, what = c("z", "deviations"),
  cluster_rows = TRUE, cluster_cols = TRUE,
  sample_annotation = colData(object), show_rownames = FALSE,
  show_colnames = FALSE, ...)
}
\arguments{
\item{object}{}

\item{what}{"z" or "deviations"}

\item{cluster_rows}{cluster rows?}

\item{cluster_cols}{cluster columns?}

\item{sample_annotation}{annotation of sample}

\item{show_rownames}{show row names?}

\item{show_colnames}{show column names?}
}
\value{
pheatmap output
}
\description{
plot heatmap of deviations
}

