% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deviations_tsne}
\alias{plot_deviations_tsne}
\title{plot_deviations_tsne}
\usage{
plot_deviations_tsne(object, tsne, var_df = NULL, annotation_column = NULL,
  motif = NULL, shiny = interactive())
}
\arguments{
\item{object}{deviations result object}

\item{tsne}{result from \code{\link{deviations_tsne}}}

\item{var_df}{variability result}

\item{annotation_column}{column name for colData(object) of annotation to be used
for coloring samples}
}
\value{
shiny widget
}
\description{
plot_deviations_tsne
}

