% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{plot_kmer_mismatch}
\alias{plot_kmer_mismatch}
\title{plot_kmer_mismatch}
\usage{
plot_kmer_mismatch(kmer, cov_mat, pval = 0.01)
}
\arguments{
\item{kmer}{kmer, e.g. 'AAAAAAA'}

\item{cov_mat}{result from \code{\link{deviations_covariability}}}

\item{pval}{p value threshold}
}
\value{
A plot
}
\description{
plot_kmer_mismatch
}

