% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResultSet.R
\docType{methods}
\name{plot_variability}
\alias{plot_variability}
\alias{plot_variability,deviationResultSet-method}
\title{plot_variability}
\usage{
plot_variability(object, ...)

\S4method{plot_variability}{deviationResultSet}(object, top = 3,
  xlab = "Sorted TFs")
}
\arguments{
\item{object}{deviationResultSet object}

\item{top}{number of top results to label}

\item{xlab}{label for x-axis (default is "sorted bins")}
}
\value{
ggplot
}
\description{
Plots variability for deviationResultSet
}
\seealso{
\code{\link{metric}}, \code{\link{variability}}
}

