% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{plotKmerMismatch}
\alias{plotKmerMismatch}
\title{plotKmerMismatch}
\usage{
plotKmerMismatch(kmer, cov_mat, pval = 0.01)
}
\arguments{
\item{kmer}{kmer, e.g. 'AAAAAAA'}

\item{cov_mat}{result from \code{\link{deviationsCovariability}}}

\item{pval}{p value threshold}
}
\value{
A plot
}
\description{
plotKmerMismatch
}
