% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm_dist.R
\name{pwm_distance}
\alias{pwm_distance}
\title{pwm_distance}
\usage{
pwm_distance(x, y = NULL, min_overlap = 5)
}
\arguments{
\item{x}{Feature 1}

\item{y}{Feature 2}

\item{min_overlap}{}
}
\value{
A distance metric between the PWM
}
\description{
pwm_distance
}

