% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peaks.R
\name{read_peaks}
\alias{read_peaks}
\title{read_peaks}
\usage{
read_peaks(filename, extra_cols = c())
}
\arguments{
\item{filename}{filename of bed file}

\item{extra_cols}{extra columns to read in beyond first three}
}
\value{
GenomicRanges containing peaks in bed file
}
\description{
Read in peaks from a bed file.
}
\details{
As in standard definition of bed file, first column is assumed to be chromosome,
second is assumed to be start of peak (0-based), and third is assumed to be end of peak (1-based).
Extra columns can be added as metadata or strand information if provided, but the user must
indicate column index and name using named vector for extra_cols
}

