% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{readNarrowpeaks}
\alias{readNarrowpeaks}
\title{readNarrowpeaks}
\usage{
readNarrowpeaks(filename, width = 500, non_overlapping = TRUE)
}
\arguments{
\item{filename}{filename}

\item{width}{desired width of peaks}

\item{non_overlapping}{remove overlapping peaks}
}
\value{
\code{\link[GenomicRanges]{GRanges-class}}
}
\description{
Reads in peaks in narrowpeaks format, as output by macs2. Uses summit as 
center of peak, and makes peak the given 'width'. By default removes 
overlapping peaks to get set of peaks with no overlaps
}
