% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peaks.R
\name{reduce_peaks}
\alias{reduce_peaks}
\title{reduce_peaks}
\usage{
reduce_peaks(counts_mat)
}
\arguments{
\item{counts_mat}{\code{\link{fragmentCounts}}}
}
\value{
\code{\link{fragmentCounts}} with only non-overlapping peaks
}
\description{
Function to eliminate overlaping peaks from counts_mat
}
\details{
This function will look for overlapping peaks and remove overlapping peaks
that have fewer fragments than the peaks they overlap
}

