% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResultSet.R
\name{subset_by_variability}
\alias{subset_by_variability}
\title{subset_by_variability}
\usage{
subset_by_variability(object, by = c("pvalue", "top", "metric", "bound"),
  cutoff = 0.05, adjusted = TRUE)
}
\arguments{
\item{object}{\code{\link{deviationResultSet}} object}

\item{by}{choice of "pvalue", "top", "metric", or "bound".  See Details.}

\item{cutoff}{-- cutoff for top sites. See Details.}

\item{adjusted}{if by is set to "pvalue", adjust for multiple comparisons?}
}
\value{
\code{\link{deviationResultSet}} object
}
\description{
function to isolate results with high variability
}
\details{
Subsetting by variability can be done in sevarl ways, determined by
the options "by" and "cutoff".  If "by" is set to "pvalue" (the default),
then the "cutoff" is an upper bound cutoff for the pvalue.  If "by" is set to
"top", then the top K most variable are returned, where K is set by "cutoff".
If "by" is "metric" then subsetting is based on the value of the variability
metric, and "cutoff" is a lower bound on that metric.  If "by" is "bound" then
subsetting is based on the value of the lower bound for the variability metric,
and "cutoff" is the lower bound for the lower bounds of the variability metric.
}

