% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{variability}
\alias{variability}
\alias{variability,deviationResult-method}
\alias{variability,deviationResultSet-method}
\title{variability}
\usage{
variability(object, ...)

\S4method{variability}{deviationResult}(object)

\S4method{variability}{deviationResultSet}(object)
}
\arguments{
\item{object}{either \code{\link{deviationResult}} or \code{\link{deviationResultSet}} object}
}
\description{
Accessor for variability metric in deviationResult or deviationResultSet.
Use \code{\link{metric}} to find metric
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns single numeric value representing the variability metric for that result

\item \code{deviationResultSet}: returns numeric vector of variability metric for all results in set
}}
\seealso{
\code{\link{metric}}, \code{\link{variability_bounds}}
}

