% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{variability_bounds}
\alias{variability_bounds}
\alias{variability_bounds,deviationResult-method}
\alias{variability_bounds,deviationResultSet-method}
\title{variability_bounds}
\usage{
variability_bounds(object, ...)

\S4method{variability_bounds}{deviationResult}(object)

\S4method{variability_bounds}{deviationResultSet}(object, lower = T)
}
\arguments{
\item{object}{either deviationResult or deviationResultSet object}

\item{lower}{return lowerbounds?  only applicable if object is deviationResultSet}

\item{lower}{return lower bound?  if FALSE, returns upper bound}
}
\description{
Accessor for variability error in deviationResult or deviationResultSet.
variability error is computed by bootstrapping the samples and determining a 95% confidence interval
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns numeric vector of length two, with first
value the lower bound and second value the upper bound

\item \code{deviationResultSet}: returns only lower or upper bound (based on lower argument) for all results in set
}}
\seealso{
\code{\link{metric}}, \code{\link{variability}}
}

