% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_synergy.R
\name{variability_synergy}
\alias{variability_synergy}
\title{variability_synergy}
\usage{
variability_synergy(object, annotations, background_peaks = NULL,
  variabilities = NULL, expectation = NULL, nbg = 25)
}
\arguments{
\item{object}{result from compute_deviations}

\item{annotations}{SummarizedExperiment of annotation matches}

\item{background_peaks}{optional, matrix of background peaks}

\item{variabilities}{optional, variabilities computed from compute_variability}

\item{expectation}{optional, expected fraction of reads per peak, as computed
by compute_expectations}

\item{nbg}{number of background iterations}
}
\value{
synergy matrix
}
\description{
variability_synergy
}
\details{
should only be run on small number of motifs/kmers/peaksets (very slow!)
}

