% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_peaks.R
\name{add_gc_bias}
\alias{add_gc_bias}
\title{add_gc_bias}
\usage{
add_gc_bias(counts_mat,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{counts_mat}{SummarizedExperiment}

\item{genome}{BSgenome object}

\item{peaks}{GenomicRanges}
}
\description{
Computes GC content for peaks
}

