% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_peaks.R
\docType{methods}
\name{add_gc_bias}
\alias{add_gc_bias}
\alias{add_gc_bias,RangedSummarizedExperiment-method}
\alias{add_gc_bias,SummarizedExperiment-method}
\title{add_gc_bias}
\usage{
add_gc_bias(object, ...)

\S4method{add_gc_bias}{RangedSummarizedExperiment}(object,
  genome = BSgenome.Hsapiens.UCSC.hg19)

\S4method{add_gc_bias}{SummarizedExperiment}(object, peaks,
  genome = BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{object}{(Ranged)SummarizedExperiment}

\item{...}{additional arguments}

\item{genome}{BSgenome object, by defualt hg19}

\item{peaks}{GenomicRanges with peaks, needed if object is 
SummarizedExperiment and not RangedSummarizedExperiment}
}
\value{
(Ranged)SummarizedExperiment object with new column in row metadata
with the gc content of the peak in question
}
\description{
Computes GC content for peaks
}
\section{Methods (by class)}{
\itemize{
\item \code{RangedSummarizedExperiment}: method for RangedSummarizedExperiment

\item \code{SummarizedExperiment}: method for SummarizedExperiment
}}

\examples{

data(example_counts, package = "chromVAR")
# show example on small part of data 
subset_counts <- example_counts[1:500,]
example_counts <- add_gc_bias(subset_counts)

}
