% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\docType{methods}
\name{annotation_matches}
\alias{annotation_matches}
\alias{annotation_matches,SummarizedExperiment-method}
\title{annotation_matches}
\usage{
annotation_matches(object)

\S4method{annotation_matches}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment with matches slot}
}
\description{
annotation_matches
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: get matches assay from SummarizedExperiment
}}

