% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backgroundPeaks.R
\docType{class}
\name{backgroundPeaks-class}
\alias{backgroundPeaks}
\alias{backgroundPeaks-class}
\title{backgroundPeaks}
\description{
backgroundPeaks is a class to store background peaks. The associated peaks are also stored
}
\section{Slots}{

\describe{
\item{\code{peaks}}{}

\item{\code{background_peaks}}{}
}}

