% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{package}
\name{chromVAR}
\alias{chromVAR}
\alias{chromVAR-package}
\title{chromVAR: A package for computing variability across sets of peaks.}
\description{
The package provides three categories of important functions:
functions!
}
\section{Functions}{

The functions ...
}

\section{Classes}{

the classes ...
}

