% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{package}
\name{chromVAR}
\alias{chromVAR}
\alias{chromVAR-package}
\title{chromVAR: A package for computing variability across sets of peaks.}
\description{
Determine variation in chromatin accessibility across sets of
annotations or peaks. Designed primarily for single-cell or sparse chromatin
accessibility, e.g. from scATAC-seq or sparse ATAC or DNAse-seq experiments.
}
