% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{chromVAR_theme}
\alias{chromVAR_theme}
\title{chromVAR_theme}
\usage{
chromVAR_theme(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
ggplot2 theme
}
\description{
theme for use with ggplot2, used by chromVAR plotting functions
}
\examples{
p <- ggplot2::qplot(1:3,1:3) + chromVAR_theme(18)
}
\author{
Alicia Schep
}
