% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{chromVARDeviations-class}
\alias{chromVARDeviations-class}
\title{chromVARDeviations}
\description{
Class for storing results from \code{\link{computeDeviations}} function.
}
\details{
This class inherits from 
\code{\link[SummarizedExperiment]{SummarizedExperiment}}, and most methods 
for that class should work for objects of this class as well. Additionally, 
two accessor functions are defined for extracting bias corrected deviations 
(\code{\link{deviations}}) and deviation Z-scores 
(\code{\link{deviationScores}})
}
