% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{compute_deviations}
\alias{compute_deviations}
\title{compute_deviations}
\usage{
compute_deviations(object, annotations = NULL, background_peaks = NULL,
  expectation = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{annotations}{SummarizedExperiment}

\item{background_peaks}{background peaks matrix}

\item{expectation}{expectations computed using \code{\link{compute_expectations}}}
}
\value{
list with two elements: 1) z: matrix with deviation z-scores,
2) dev: matrix with normalized deviations
}
\description{
Computes deviations across sets of annotations
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\seealso{
\code{\link{compute_variability}}, \code{\link{plot_variability}}
}

