% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{compute_deviations}
\alias{compute_deviations}
\alias{compute_deviations,MatrixOrmatrix,MatrixOrmatrix-method}
\alias{compute_deviations,MatrixOrmatrix,SummarizedExperiment-method}
\alias{compute_deviations,MatrixOrmatrix,list-method}
\alias{compute_deviations,MatrixOrmatrix,missingOrNULL-method}
\alias{compute_deviations,SummarizedExperiment,MatrixOrmatrix-method}
\alias{compute_deviations,SummarizedExperiment,SummarizedExperiment-method}
\alias{compute_deviations,SummarizedExperiment,list-method}
\alias{compute_deviations,SummarizedExperiment,missingOrNULL-method}
\title{compute_deviations}
\usage{
compute_deviations(object, annotations, ...)


  \S4method{compute_deviations}{SummarizedExperiment,SummarizedExperiment}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

\S4method{compute_deviations}{SummarizedExperiment,MatrixOrmatrix}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

\S4method{compute_deviations}{SummarizedExperiment,list}(object, annotations,
  background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

\S4method{compute_deviations}{SummarizedExperiment,missingOrNULL}(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

\S4method{compute_deviations}{MatrixOrmatrix,SummarizedExperiment}(object,
  annotations, background_peaks, expectation = compute_expectations(object))

\S4method{compute_deviations}{MatrixOrmatrix,MatrixOrmatrix}(object,
  annotations, background_peaks, expectation = compute_expectations(object))

\S4method{compute_deviations}{MatrixOrmatrix,list}(object, annotations,
  background_peaks, expectation = compute_expectations(object))

\S4method{compute_deviations}{MatrixOrmatrix,missingOrNULL}(object, annotations,
  background_peaks, expectation = compute_expectations(object))
}
\arguments{
\item{object}{chromVARCounts object}

\item{annotations}{chromVARAnnotations object}

\item{...}{additional arguments}

\item{background_peaks}{(optional) background peaks matrix computed using 
\code{\link{get_background_peaks}}, computed internally with default
paramaters if not provided}

\item{expectation}{(optional) expectations computed using
\code{\link{compute_expectations}}, computed automatically if not provided}
}
\value{
\code{\link{chromVARDeviations-class}}, which inherits from 
SummarizedExperiment, and has two assays: deviations and deviation scores.
}
\description{
Computes deviations in chromatin accessibility across sets of annotations
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SummarizedExperiment,annotations = SummarizedExperiment}: object and annotations are SummarizedExperiment

\item \code{object = SummarizedExperiment,annotations = MatrixOrmatrix}: object is SummarizedExperiment, 
annotations are Matrix

\item \code{object = SummarizedExperiment,annotations = list}: object is SummarizedExperiment, 
annotations are list

\item \code{object = SummarizedExperiment,annotations = missingOrNULL}: object is SummarizedExperiment, 
annotations are missing

\item \code{object = MatrixOrmatrix,annotations = SummarizedExperiment}: object and annotations are SummarizedExperiment

\item \code{object = MatrixOrmatrix,annotations = MatrixOrmatrix}: object is SummarizedExperiment, 
annotations are Matrix

\item \code{object = MatrixOrmatrix,annotations = list}: object is SummarizedExperiment, 
annotations are list

\item \code{object = MatrixOrmatrix,annotations = missingOrNULL}: object is SummarizedExperiment, 
annotations are missing
}}
\seealso{
\code{\link{compute_variability}}, \code{\link{plot_variability}}
}

