% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_deviations}
\alias{compute_deviations}
\title{compute_deviations}
\usage{
compute_deviations(peak_set, counts_mat, niterations = 50,
  metric = c("z-score", "old"), intermediate_results = FALSE)
}
\arguments{
\item{peak_set}{vector of indices for peaks in set}

\item{counts_mat}{fragmentCounts object}

\item{niterations}{number of background sets to sample}

\item{metric}{which metric to use?  default is z-score}
}
\value{
\code{\link{deviationResult}}
}
\description{
Computes deviations across samples for a set of annotations
}
\seealso{
\code{\link{deviationResult}}, \code{\link{compute_variability}}
}

