% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{compute_deviations}
\alias{compute_deviations}
\title{compute_deviations}
\usage{
compute_deviations(counts_mat, background_peaks, peak_indices = NULL,
  expectation = NULL, norm = TRUE)
}
\arguments{
\item{counts_mat}{matrix with counts}

\item{background_peaks}{background peaks matrix}

\item{peak_indices}{list of indices representing different sets of peaks}

\item{expectation}{expectations computed using \code{\link{compute_expectations}}}
}
\value{
list with two elements: 1) z: matrix with deviation z-scores,
2) fc: matrix with deviation log2 fold-changes
}
\description{
Computes deviations across sets of annotations
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\seealso{
\code{\link{compute_variability}}, \code{\link{plot_variability}}
}

