% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{compute_expectations}
\alias{compute_expectations}
\title{compute_expectations}
\usage{
compute_expectations(object, norm = FALSE, annotation = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{norm}{weight all samples equally?}

\item{annotation}{an annotation vector, optional}
}
\value{
vector with expected fraction of reads per peak
}
\description{
compute_expectations
}
\details{
By default, this function will compute the expected fraction of reads
per peak as the the total fragments per peak across all samples divided by total
reads in peaks in all samples. Optionally, norm can be set to TRUE and then the
expectation will be the average fraction of reads in a peak across the cells.
This is not recommended for single cell applications as cells with very few reads
will have a large impact.  Another option is to give a vector of annotations, in which
case the expectation will be the average fraction of reads per peak within each annotation.
If annotation vector is provided and norm is set to TRUE then within each annotation
the fraction of reads per peak is the average fraction of reads per peak in each
sample.  Otherwise, the within annotation fraction of reads per peak is based on the
reads per peak within the sample divided by the total reads within each sample.
}

