% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_variability}
\alias{compute_variability}
\title{compute_variability}
\usage{
compute_variability(motif_indices, counts_mat, niterations = 50,
  metric = c("z-score", "old"), count = NULL)
}
\arguments{
\item{motif_indices}{list of indices representing different sets of peaks}

\item{counts_mat}{fragmentCounts object}

\item{niterations}{number of background sets to sample}

\item{metric}{which metric to use? default is z-score}

\item{count}{is input data count data?  will determine based on data if not given}
}
\value{
\code{\link{deviationResultSet}}
}
\description{
Computes variability across sets of annotations
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\seealso{
\code{\link{deviationResultSet}}, \code{\link{variability}}, \code{\link{plot_variability}}
}

