% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_variability}
\alias{compute_variability}
\title{compute_variability}
\usage{
compute_variability(motif_indices, counts_mat, niterations = 50,
  metric = c("z-score", "old"), BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{motif_indices}{list of indices representing different sets of peaks}

\item{counts_mat}{fragmentCounts object}

\item{niterations}{number of background sets to sample}

\item{metric}{which metric to use? default is z-score}

\item{BPPARAM}{multi-processing argument to pass to \code{\link[BiocParallel]{bplapply}}}
}
\value{
\code{\link{deviationResultSet}}
}
\description{
Computes variability across sets of annotations
}
\seealso{
\code{\link{deviationResultSet}}, \code{\link{variability}}, \code{\link{plot_variability}}
}

