% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_variability}
\alias{compute_variability}
\title{compute_variability}
\usage{
compute_variability(object, bootstrap_error = TRUE,
  bootstrap_samples = 1000, bootstrap_quantiles = c(0.025, 0.975),
  na.rm = TRUE)
}
\arguments{
\item{object}{output from \code{\link{compute_deviations}}}

\item{bootstrap_error}{compute bootstrap confidence interval}

\item{bootstrap_samples}{number of bootstrap samples to take}

\item{bootstrap_quantiles}{quantiles for bootstrap}

\item{na.rm}{remove NAs? default is true}
}
\value{
data.frame with columns for name, variability, bootstrap lower bound,
bootstrap upper bound, raw p value, adjust p value.
}
\description{
function to compute overall variability of motif sets across samples
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
variability <- compute_variability(dev)                        
}
