% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_variability.R
\name{compute_variability}
\alias{compute_variability}
\title{compute_variability}
\usage{
compute_variability(deviations, bootstrap_error = TRUE,
  bootstrap_samples = 1000, bootstrap_quantiles = c(0.025, 0.975))
}
\arguments{
\item{deviations}{output from \code{\link{compute_deviations}}}

\item{bootstrap_error}{compute bootstrap confidence interval}

\item{bootstrap_samples}{number of bootstrap samples to take}

\item{bootstrap_quantiles}{quantiles for bootstrap}
}
\description{
function to compute overall variability of motif sets across samples
}

