% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{deviation_scores}
\alias{deviation_scores}
\title{deviation_scores}
\usage{
deviation_scores(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\description{
Accessor for deviation Z-scores from \code{\link{chromVARDeviations-class}} object
}

