% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResultSet.R
\docType{class}
\name{deviationResultSet-class}
\alias{deviationResultSet}
\alias{deviationResultSet-class}
\title{A class to store deviationResult objects}
\description{
A class to store deviationResult objects
}
\seealso{
\code{\link{metric}}, \code{\link{variability}}, \code{\link{variability_bounds}}
}

