% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\name{deviations}
\alias{deviations}
\title{deviations}
\usage{
deviations(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\description{
Accessor for bias corrected deviations from \code{\link{chromVARDeviations-class}} object
}

