% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{deviations}
\alias{deviations}
\alias{deviations,chromVARDeviations-method}
\alias{deviations,chromVARDeviations-method}
\title{deviations}
\usage{
deviations(object)

\S4method{deviations}{chromVARDeviations}(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\value{
matrix of bias corrected deviations
}
\description{
Accessor for bias corrected deviations from \code{\link{chromVARDeviations-class}} object
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
bias_corrected_deviations <- deviations(dev)
}
\author{
Alicia Schep
}
