% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{deviations}
\alias{deviations}
\alias{deviations,deviationResult-method}
\alias{deviations,deviationResultSet-method}
\title{deviations}
\usage{
deviations(object, ...)

\S4method{deviations}{deviationResult}(object)

\S4method{deviations}{deviationResultSet}(object)
}
\arguments{
\item{object}{either \code{\link{deviationResult}} or \code{\link{deviationResultSet}}}
}
\description{
Accessor for deviations z-score in deviationResult.
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns vector of deviations for samples

\item \code{deviationResultSet}: returns matrix of deviations, with columns representing
samples and rows representing annotation sets
}}
\seealso{
\code{\link{foldchange}}, \code{\link{variability_bounds}}
}

