% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{deviations_covariability}
\alias{deviations_covariability}
\title{deviations_covariability}
\usage{
deviations_covariability(object)
}
\arguments{
\item{object}{deviations result}
}
\value{
'covariability' matrix
}
\description{
deviations_covariability
}
\details{
Returns the 'covariability' between motifs/kmers/peaksets. 
Covariability' is defined as covariance between Z-scores divided by variance 
of Z-scores for one motif/kmer/peakset (the row).
}

