% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{deviations_tsne}
\alias{deviations_tsne}
\title{deviations_tsne}
\usage{
deviations_tsne(object, threshold = 1.5, perplexity = if (what == "samples")
  30 else 8, max_iter = 1000, theta = 0.5, what = c("samples",
  "annotations"), shiny = FALSE)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{variability threshold -- use only deviations with 
variability greater than threshold}

\item{perplexity}{perplexity parameter for tsne}

\item{max_iter}{max iterations parameter for tsne}

\item{theta}{theta parameter for tsne}

\item{what}{tsne for similarity of samples or annotations?}

\item{shiny}{load a shiny widget that enable you to explore perplexity and 
variability threshold parameter?}
}
\value{
data.frame with two columns for the two dimensions of tSNE output
}
\description{
Perform tsne using bias corrected deviations to visualize either cell/sample
similarity or motif/kmer/annotation similarity
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
                         
# Not run                         
\dontrun{sample_dist <- deviations_tsne(dev, threshold = 1)}  
# setting very low variabilitiy threshold because this is mini data set
# Use plot_variability to get a sense of an appropriate threshold
# This gives an error with this mini data set as there too many NA values
}
\author{
Alicia Schep
}
