% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{deviations_tsne}
\alias{deviations_tsne}
\title{deviations_tsne}
\usage{
deviations_tsne(object, threshold = 1.5, perplexity = 30,
  shiny = interactive(), ...)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{variability threshold -- use only deviations with variability greater than threshold}

\item{perplexity}{perplexity parameter for tsne}

\item{shiny}{output shiny gadget for choosing parameters & visualizing results}

\item{...}{}
}
\value{
list with three elements: threshold used, perplexity used, tsne results 
from \code{\link[Rtsne]{Rtsne}}
}
\description{
deviations_tsne
}

