% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{deviationScores}
\alias{deviationScores}
\alias{deviationScores,chromVARDeviations-method}
\alias{deviationScores,chromVARDeviations-method}
\title{deviationScores}
\usage{
deviationScores(object)

\S4method{deviationScores}{chromVARDeviations}(object)
}
\arguments{
\item{object}{chromVARDeviations object}
}
\value{
The deviationScores and deviations accessors both return matrices.

matrix of deviation Z-scores
}
\description{
Accessor for deviation Z-scores from 
\code{\link{chromVARDeviations-class}} object
}
\examples{
# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")
scores <- deviationScores(mini_dev)
}
\author{
Alicia Schep
}
