% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{differential_deviations}
\alias{differential_deviations}
\title{differential_deviations}
\usage{
differential_deviations(object, groups, alternative = c("two.sided", "less",
  "greater"), parametric = TRUE)
}
\arguments{
\item{object}{chromVARDeviations object}

\item{groups}{either vector of groups or name of column in colData of object with grouop information}

\item{alternative}{only used if there are two groups -- two.sided or one sided test}

\item{parametric}{use parametric test. alternatively will use kruskal wallace}
}
\value{
data.frame with p value and adjusted p value
}
\description{
Function to see whether deviations differ between groups
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
library(BSgenome.Hsapiens.UCSC.hg19)
motif_ix <- match_motifs(motifs, mini_counts, 
                         genome = BSgenome.Hsapiens.UCSC.hg19)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
difdev <- differential_deviations(dev, "Cell_Type")
}
\author{
Alicia Schep
}
