% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_tests.R
\name{differential_variability}
\alias{differential_variability}
\title{differential_variability}
\usage{
differential_variability(object, groups, parametric = TRUE)
}
\arguments{
\item{object}{chromVARDeviations object}

\item{groups}{either vector of groups or name of column in colData of object with grouop information}

\item{parametric}{use parametric test. alternatively will use kruskal wallace}
}
\value{
data.frame with p value and adjusted p value
}
\description{
Function to determine whether groups differ in variability
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
difvar <- differential_variability(dev, "Cell_Type")
}
\author{
Alicia Schep
}
