% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_peaks}
\alias{filter_peaks}
\title{filter_peaks}
\usage{
filter_peaks(object, min_fragments_per_peak = 1, non_overlapping = TRUE,
  ix_return = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment with matrix of fragment counts per peak 
per sample, as computed by \code{\link{get_counts}}}

\item{min_fragments_per_peak}{minimum number of fragmints in peaks across all 
samples}

\item{non_overlapping}{reduce peak set to non-overlapping peaks, see details}

\item{ix_return}{return indices of peaks to keep instead of subsetted counts object}
}
\value{
vector of indices, representing peaks that should be kept
}
\description{
function to get indices of peaks that pass filters
}
\details{
if non_overlapping is set to true, when peaks overlap the overlapping
peak with lower counts is removed
}
\examples{
data(example_counts, package = "chromVAR")

counts_filtered <- filter_samples(example_counts, min_depth = 1500,
                                  min_in_peaks = 0.15, shiny = FALSE)
counts_filtered <- filter_peaks(example_counts)
}
\seealso{
\code{\link{get_peaks}},  \code{\link{filter_samples}},
\code{\link{get_counts}}
}
\author{
Alicia Schep
}
