% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_peaks}
\alias{filter_peaks}
\title{filter_peaks}
\usage{
filter_peaks(counts_mat, min_fragments_per_peak = 1, non_overlapping = TRUE,
  ix_return = FALSE)
}
\arguments{
\item{counts_mat}{SummarizedExperiment with matrix of fragment counts per peak per sample, as computed 
by \code{\link{get_counts}}
\code{\link{read_peaks}}}

\item{min_fragments_per_peak}{minimum number of fragmints in peaks across all samples}

\item{non_overlapping}{reduce peak set to non-overlapping peaks, see details}
}
\value{
vector of indices, representing peaks that should be kept
}
\description{
function to get indices of peaks that pass filters
}
\details{
if non_overlapping is set to true, when peaks overlap the overlapping peak with lower counts is removed
}
\seealso{
\code{\link{get_peaks}},  \code{\link{get_peaks}}, \code{\link{filter_samples}},
\code{\link{get_counts}}
}

