% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_samples}
\alias{filter_samples}
\title{filter_samples}
\usage{
filter_samples(counts_mat, depths, min_in_peaks = NULL, min_depth = NULL,
  plot = TRUE)
}
\arguments{
\item{counts_mat}{matrix of fragment counts per peak per sample, as computed
by \code{\link{getFragmentCounts}}}

\item{depths}{vector of sequencing depth per samples, as computed by
\code{\link{getSampleDepths}}}

\item{min_in_peaks}{minimum fraction of samples within peaks}

\item{min_depth}{minimum library size}

\item{plot}{show plot?}
}
\value{
indices of samples to keep
}
\description{
function to get indices of samples that pass filtters
}
\details{
If unspecified, min_in_peaks and min_depth cutoffs will be estimated based on data.
min_in_peaks is set to 0.5 times the median proportion of fragments in peaks.  min_depth is
set to the maximum of 500 or 10% of the median library size.
}
\seealso{
\code{\link{get_counts}},  \code{\link{get_inputs}}, \code{\link{filter_peaks}}
}

