% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{filterFragmentCounts}
\alias{filterFragmentCounts}
\title{Filter Fragment Counts by quality}
\usage{
filterFragmentCounts(counts_mat, min_in_peaks = 0.25, min_fragments = 5000)
}
\arguments{
\item{counts_mat}{fragmentCounts object}

\item{min_in_peaks}{minimum fraction of fragments within peaks (default = 0.25)}

\item{min_fragments}{minimum fragments in peaks (default = 5000)}
}
\description{
Filter Fragment Counts by quality
}

