% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tfdev.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{tfdev: A package for computing variabilitiy across sets of peaks.}
\description{
The package provides three categories of important functions:
functions!
}
\section{Functions}{

The functions ...
}

\section{Classes}{

the classes ...
}

