% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{get_fragments_per_peak}
\alias{get_fragments_per_peak}
\alias{get_fragments_per_peak,SummarizedExperiment-method}
\alias{get_fragments_per_peak,MatrixOrmatrix-method}
\title{get_fragments_per_peak}
\usage{
get_fragments_per_peak(object)

\S4method{get_fragments_per_peak}{SummarizedExperiment}(object)

\S4method{get_fragments_per_peak}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
vector with sum across rows of counts assay within chromVARCounts
}
\description{
get_fragments_per_peak
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")

frags_per_peak <- get_fragments_per_peak(mini_counts)
}
\seealso{
\code{\link{get_fragments_per_sample}},
\code{\link{get_total_fragments}}
}
