% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_fragments_per_sample}
\alias{get_fragments_per_sample}
\title{get_fragments_per_peak}
\usage{
get_fragments_per_sample(input)
}
\arguments{
\item{input}{SummarizedExperiment object}
}
\value{
vector with sum across columns of counts assay within SummarizedExperiment
}
\description{
get_fragments_per_peak
}

