% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{get_fragments_per_sample}
\alias{get_fragments_per_sample}
\alias{get_fragments_per_sample,SummarizedExperiment-method}
\alias{get_fragments_per_sample,MatrixOrmatrix-method}
\title{get_fragments_per_sample}
\usage{
get_fragments_per_sample(object)

\S4method{get_fragments_per_sample}{SummarizedExperiment}(object)

\S4method{get_fragments_per_sample}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
vector with sum across columns of counts assay within chromVARCounts
}
\description{
get_fragments_per_sample
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
frags_per_sample <- get_fragments_per_sample(mini_counts)
}
\seealso{
\code{\link{get_fragments_per_peak}},
\code{\link{get_total_fragments}}
}
