% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifs.R
\name{get_motif_indices}
\alias{get_motif_indices}
\title{get_motif_indices}
\usage{
get_motif_indices(motifs, peaks,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  p.cutoff = 5e-05)
}
\arguments{
\item{motifs}{\code{\link[TFBSTools]{PFMatrixList}} or \code{\link[TFBSTools]{PWMatrixList}}}

\item{peaks}{\code{\link[GenomicRanges]{GenomicRanges}} or \code{\link{fragmentCounts}}}

\item{genome}{\code{\link[BSgenome]{BSgenome}} object}

\item{p.cutoff}{default is 0.00005}
}
\value{
A list with a vector of peak indices for each motif.
}
\description{
Function to get indices of peaks that contain motifs
}

