% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_sample_correlation}
\alias{get_sample_correlation}
\title{get_sample_correlation}
\usage{
get_sample_correlation(object, threshold = 1.5)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{threshold for variability}
}
\value{
correlation matrix between samples
}
\description{
Get correlation between samples based on bias corrected deviations
}
\details{
This function will compute the correlation between samples based on 
the normalized deviations. It will first remove correlated motifs/peak sets. 
Then the pearson correlation coefficient will be computed and returned.
}
\author{
Alicia Schep
}
