% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{get_sample_depths}
\alias{get_sample_depths}
\title{get_sample_depths}
\usage{
get_sample_depths(bams, paired = TRUE, by_rg = FALSE)
}
\arguments{
\item{bams}{filenames for bam file with aligned reads}

\item{paired}{paired end data?}

\item{by_rg}{use RG tags to separate groups?}
}
\value{
numeric vector
}
\description{
makes vector of read depths in bam files or RG groups within bam files
}
\seealso{
\code{\link{get_counts}},  \code{\link{get_inputs}}, \code{\link{filter_samples}}
}

