% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_sample_distance}
\alias{get_sample_distance}
\title{get_sample_distance}
\usage{
get_sample_distance(object, threshold = 1.5, initial_dims = 50,
  distance_function = dist)
}
\arguments{
\item{object}{deviations result}

\item{threshold}{threshold for variability}

\item{initial_dims}{initial dimentions for preliminary dimensionality reduction via pca}

\item{distance_function}{distance function to use}
}
\value{
dist object for distance between samples
}
\description{
get_sample_distance
}
\details{
This function will compute the distance between samples based on the 
normalized deviations.  It will first remove correlated motifs / peak sets.  Then
the dimensionality will be further reduced via PCA if the number of dimensions
exceeds initial_dims.  Then  the supplied distance_function will be use.
}

