% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_total_fragments}
\alias{get_total_fragments}
\title{get_total_fragments}
\usage{
get_total_fragments(input)
}
\arguments{
\item{input}{SummarizedExperiment object}
}
\value{
sum of all counts within input
}
\description{
get_total_fragments
}

