% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backgroundPeaks.R
\name{getBackgroundPeakSets}
\alias{getBackgroundPeakSets}
\title{getBackgroundPeakSets}
\usage{
getBackgroundPeakSets(counts_mat, niterations = 50, window = 1000,
  BPPARAM = BiocParallel::bpparam(), for_set = TRUE)
}
\arguments{
\item{counts_mat}{a fragmentCounts object}

\item{niterations}{number of background peaks to sample}

\item{window}{window size around peak in which to sample a background peak}

\item{BPPARAM}{multiprocessing parameter for use by BiocParallel}

\item{for_set}{computing background peaks for a set of peaks? affect whether sampling is done with replacement}
}
\value{
backgroundPeaks object
}
\description{
Function to get a set of background peaks for each peak based on GC content and # of counts
across all samples
}
\details{
Background peaks are chosen by finding the window nearest neighbors to a peak in terms of
GC content and # of fragments across samples using the Mahalanobis distance.  From those nearest
neighbors, niterations peaks are sampled from that background.
}

