% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{getFragmentCountsByRG}
\alias{getFragmentCountsByRG}
\title{getFragmentCountsByRG}
\usage{
getFragmentCountsByRG(bam, peaks)
}
\arguments{
\item{bam}{filename for bam file with aligned reads}

\item{peaks}{GRanges object with peaks}
}
\value{
\code{\link{fragmentCounts}} object
}
\description{
make fragmentCounts object using bam file with RG tags
}
\seealso{
\code{\link{getFragmentCounts}}, \code{\link{fragmentCounts}}, \code{\link{filterFragmentCounts}}
}

