% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{getFragmentsPerPeak}
\alias{getFragmentsPerPeak}
\alias{getFragmentsPerPeak,SummarizedExperiment-method}
\alias{getFragmentsPerPeak,MatrixOrmatrix-method}
\title{getFragmentsPerPeak}
\usage{
getFragmentsPerPeak(object)

\S4method{getFragmentsPerPeak}{SummarizedExperiment}(object)

\S4method{getFragmentsPerPeak}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
vector with sum across rows of counts assay within chromVARCounts
}
\description{
getFragmentsPerPeak
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")

frags_per_peak <- getFragmentsPerPeak(mini_counts)
}
\seealso{
\code{\link{getFragmentsPerSample}},
\code{\link{getTotalFragments}}
}
