% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{getFragmentsPerSample}
\alias{getFragmentsPerSample}
\alias{getFragmentsPerSample,SummarizedExperiment-method}
\alias{getFragmentsPerSample,MatrixOrmatrix-method}
\title{getFragmentsPerSample}
\usage{
getFragmentsPerSample(object)

\S4method{getFragmentsPerSample}{SummarizedExperiment}(object)

\S4method{getFragmentsPerSample}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
vector with sum across columns of counts assay within chromVARCounts
}
\description{
getFragmentsPerSample
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
frags_per_sample <- getFragmentsPerSample(mini_counts)
}
\seealso{
\code{\link{getFragmentsPerPeak}},
\code{\link{getTotalFragments}}
}
