% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{getJasparMotifs}
\alias{getJasparMotifs}
\title{getJasparMotifs}
\usage{
getJasparMotifs(species = "Homo sapiens", collection = "CORE", ...)
}
\arguments{
\item{species}{Which species?  use eithe jaspar code or latin name. 
default is 'Homo sapiens'}

\item{collection}{Which collection to use?  default is 'CORE'}

\item{...}{additional arguments to opts for 
\code{\link[TFBSTools]{getMatrixSet}}}
}
\value{
\code{\link[TFBSTools]{PFMatrixList}}
}
\description{
Function to get motifs from JASPAR database
}
\details{
Simply a wrapper function for \code{\link[TFBSTools]{getMatrixSet}}
 that calls JASPAR2016 database using \code{\link[JASPAR2016]{JASPAR2016}}
}
\examples{

motifs <- getJasparMotifs()


}
