% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{getTotalFragments}
\alias{getTotalFragments}
\alias{getTotalFragments,SummarizedExperiment-method}
\alias{getTotalFragments,MatrixOrmatrix-method}
\title{getTotalFragments}
\usage{
getTotalFragments(object)

\S4method{getTotalFragments}{SummarizedExperiment}(object)

\S4method{getTotalFragments}{MatrixOrmatrix}(object)
}
\arguments{
\item{object}{SummarizedExperiment, matrix, or Matrix object}
}
\value{
sum of all counts within object
}
\description{
getTotalFragments
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment object 
with counts assay

\item \code{MatrixOrmatrix}: method for Matrix or matrix object
}}

\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
total_frags <- getTotalFragments(mini_counts)
}
\seealso{
\code{\link{getFragmentsPerSample}}, 
\code{\link{getFragmentsPerPeak}}
}
