% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pwms.R
\name{match_pwms}
\alias{match_pwms}
\title{match_pwms}
\usage{
match_pwms(pwms, subject, ...)
}
\arguments{
\item{pwms}{either \code{\link[TFBSTools]{PFMatrix}}, \code{\link[TFBSTools]{PFMatrixList}},
\code{\link[TFBSTools]{PWMatrix}}, \code{\link[TFBSTools]{PWMatrixList}}}

\item{subject}{either \code{\link[GenomicRanges]{GenomicRanges}}, \code{\link[Biostrings]{DNAStringSet}},
\code{\link[Biostrings]{DNAString}}, or character vector}

\item{genome}{BSgenome object, only used if subect is \code{\link[GenomicRanges]{GenomicRanges}}}

\item{bg}{background nucleotide frequencies. if not provided, computed from subject}

\item{out}{what to return? see details}

\item{p.cutoff}{p-value cutoff for returning motifs}

\item{w}{parameter controlling size of window for filtration; default is 7}

\item{if}{subject is not GenomicRanges, ranges to use when out is positions}
}
\description{
Find pwm matches
}
\details{
Can either return a SummarizedExperiment with just sparse matrix with values set to 1 for a match (if return == "matches"), 
a SummarizedExperiment with a matches matrix as well as matrices with the maximum motif score and total motif counts (if return == "scores"), or
a GenomicRanges or IRanges object with all the positions of matches   
or \code{\link[GenomicRanges]{GenomicRanges}} if positions
}

