% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pwms.R
\docType{methods}
\name{match_pwms}
\alias{match_pwms}
\alias{match_pwms,PFMatrix,ANY-method}
\alias{match_pwms,PFMatrixList,ANY-method}
\alias{match_pwms,PWMatrix,ANY-method}
\alias{match_pwms,PWMatrixList,DNAString-method}
\alias{match_pwms,PWMatrixList,DNAStringSet-method}
\alias{match_pwms,PWMatrixList,GenomicRanges-method}
\alias{match_pwms,PWMatrixList,RangedSummarizedExperiment-method}
\alias{match_pwms,PWMatrixList,character-method}
\title{match_pwms}
\usage{
match_pwms(pwms, subject, ...)

\S4method{match_pwms}{PWMatrixList,DNAStringSet}(pwms, subject, genome = NULL,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PWMatrixList,character}(pwms, subject, genome = NULL,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PWMatrixList,DNAString}(pwms, subject, genome = NULL,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PWMatrixList,GenomicRanges}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PWMatrixList,RangedSummarizedExperiment}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PFMatrixList,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PWMatrix,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_pwms}{PFMatrix,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)
}
\arguments{
\item{pwms}{either \code{\link[TFBSTools]{PFMatrix}}, \code{\link[TFBSTools]{PFMatrixList}},
\code{\link[TFBSTools]{PWMatrix}}, \code{\link[TFBSTools]{PWMatrixList}}}

\item{subject}{either \code{\link[GenomicRanges]{GenomicRanges}}, \code{\link[Biostrings]{DNAStringSet}},
\code{\link[Biostrings]{DNAString}}, or character vector}

\item{genome}{BSgenome object, only used if subect is \code{\link[GenomicRanges]{GenomicRanges}}}

\item{bg}{background nucleotide frequencies. if not provided, computed from subject}

\item{out}{what to return? see details}

\item{p.cutoff}{p-value cutoff for returning motifs}

\item{w}{parameter controlling size of window for filtration; default is 7}

\item{ranges}{if subject is not GenomicRanges, ranges to use when out is positions}
}
\description{
Find pwm matches
}
\details{
Can either return a SummarizedExperiment with just sparse matrix with values set to 1 for a match (if return == "matches"), 
a SummarizedExperiment with a matches matrix as well as matrices with the maximum motif score and total motif counts (if return == "scores"), or
a GenomicRanges or IRanges object with all the positions of matches   
or \code{\link[GenomicRanges]{GenomicRanges}} if positions
}
\section{Methods (by class)}{
\itemize{
\item \code{pwms = PWMatrixList,subject = DNAStringSet}: PWMatrixList/DNAStringSet

\item \code{pwms = PWMatrixList,subject = character}: PWMatrixList/character

\item \code{pwms = PWMatrixList,subject = DNAString}: PWMatrixList/DNAString

\item \code{pwms = PWMatrixList,subject = GenomicRanges}: PWMatrixList/GenomicRanges

\item \code{pwms = PWMatrixList,subject = RangedSummarizedExperiment}: PWMatrixList/RangedSummarizedExperiment

\item \code{pwms = PFMatrixList,subject = ANY}: PFMatrixList/ANY

\item \code{pwms = PWMatrix,subject = ANY}: PWMatrix/ANY

\item \code{pwms = PFMatrix,subject = ANY}: PFMatrix/ANY
}}

