% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviationResult.R, R/deviationResultSet.R
\docType{methods}
\name{metric}
\alias{metric}
\alias{metric,deviationResult-method}
\alias{metric,deviationResultSet-method}
\title{metric}
\usage{
metric(object, ...)

\S4method{metric}{deviationResult}(object)

\S4method{metric}{deviationResultSet}(object)
}
\arguments{
\item{object}{either \code{\link{deviationResult}} or \code{\link{deviationResultSet}} object}
}
\description{
Accessor for which variability metric was used.
}
\section{Methods (by class)}{
\itemize{
\item \code{deviationResult}: returns name of metric used

\item \code{deviationResultSet}: returns name of metric used
}}
\seealso{
\code{\link{variability}}
}

