% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR.R
\docType{data}
\name{mouse_pwms_v1}
\alias{mouse_pwms_v1}
\title{mouse_pwms_v1}
\usage{
data(mouse_pwms_v1)
}
\value{
\code{\link[TFBSTools]{XMatrixList}} of length 1346
}
\description{
Collection of mouse pwms
}
\examples{
data(mouse_pwms_v1)
}
\keyword{datasets}
