% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deviations}
\alias{plot_deviations}
\title{plot_deviations}
\usage{
plot_deviations(object, xlabel = NA, ylabel = NA, name = NA,
  cluster_row = TRUE, cluster_col = TRUE, cluster_row_dist = cor_dist,
  cluster_col_dist = cor_dist, sample_annotation = NULL)
}
\arguments{
\item{object}{\code{\link{deviationResultSet}} object}
}
\value{
ggplot object
}
\description{
plot_deviations
}

