% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variability}
\alias{plot_variability}
\title{plot_variability}
\usage{
plot_variability(variability, xlab = "Sorted TFs", n = 3,
  labels = variability$name, use_plotly = interactive())
}
\arguments{
\item{variability}{output from \code{\link{compute_variability}}}

\item{xlab}{label for x-axis (default is 'Sorted TFs')}

\item{n}{number of toppoints to label?}

\item{labels}{names of sets. if not given, uses rownames of variability}

\item{use_plotly}{make plot interactive (using plotly)}
}
\value{
ggplot or plotly object, depending on whether use_plotly is TRUE
}
\description{
plot variability of motifs/etc
}
\examples{
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
library(BSgenome.Hsapiens.UCSC.hg19)
motif_ix <- match_motifs(motifs, mini_counts, 
                         genome = BSgenome.Hsapiens.UCSC.hg19)

# computing deviations
dev <- compute_deviations(object = mini_counts, 
                         annotations = motif_ix)
variability <- compute_variability(dev)
var_plot <- plot_variability(variability, use_plotly = FALSE)                        
}
\author{
Alicia Schep
}
