% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variability}
\alias{plot_variability}
\title{plot_variability}
\usage{
plot_variability(variability, xlab = "Sorted TFs", n = 3, labels = NULL)
}
\arguments{
\item{variability}{output from \code{\link{compute_variability}}}

\item{xlab}{label for x-axis (default is "Sorted TFs")}

\item{n}{number of toppoints to label?}

\item{labels}{names of sets. if not given, uses rownames of variability}
}
\description{
plot variability of motifs/etc
}

