% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variability}
\alias{plot_variability}
\title{plot_variability}
\usage{
plot_variability(variability, xlab = "Sorted TFs", label_type = c("top",
  "names", "none"), label_option = NULL, only = FALSE)
}
\arguments{
\item{xlab}{label for x-axis (default is "sorted bins")}

\item{label_type}{which points to label? See Details.}

\item{label_option}{See Details.}

\item{only}{FALSE, p-value cutoff, or number of points.  See Details.}

\item{object}{\code{\link{deviationResultSet}} object}
}
\description{
plot variability of motifs/etc
}
\details{
Labelling of points is controlled by label_type and label_options.
label_type can be either "top", "names", or "none".
}

