% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDeviationsTsne}
\alias{plotDeviationsTsne}
\title{plotDeviationsTsne}
\usage{
plotDeviationsTsne(object, tsne, var_df = NULL, sample_column = NULL,
  annotation_name = NULL, shiny = interactive())
}
\arguments{
\item{object}{deviations result object}

\item{tsne}{result from \code{\link{deviationsTsne}}}

\item{var_df}{variability result}

\item{sample_column}{column name for sample data -- colData(object) -- to be
used for coloring points}

\item{annotation_name}{name of chromVAR annotation for coloring points}

\item{shiny}{return shiny app?  otherwise return static plots}
}
\value{
shiny app or plots
}
\description{
plots sample similarity tsne
}
\author{
Alicia Schep
}
