% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{rbind,chromVARDeviations-method}
\alias{rbind,chromVARDeviations-method}
\title{rbind method chromVARDeviations}
\usage{
\S4method{rbind}{chromVARDeviations}(..., deparse.level = 1)
}
\arguments{
\item{...}{chromVARDeviations object to be combined}

\item{deparse.level}{See ?base::rbind for a description of this argument.}
}
\description{
Concatenates chromVARDeviations results for different sets of annotations
}

