% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_deviations.R
\docType{methods}
\name{rbind,chromVARDeviations-method}
\alias{rbind,chromVARDeviations-method}
\title{rbind method chromVARDeviations}
\usage{
\S4method{rbind}{chromVARDeviations}(..., deparse.level = 1)
}
\arguments{
\item{...}{chromVARDeviations object to be combined}

\item{deparse.level}{See ?base::rbind for a description of this argument.}
}
\value{
chromVARDeviations object
}
\description{
Concatenates chromVARDeviations results for different sets of annotations
}
\examples{
# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")
doubledev <- rbind(mini_dev, mini_dev) #concatenate two of the same tother
}
\seealso{
\code{\link{chromVARDeviations-class}}
}
\author{
Alicia Schep
}
