% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backgroundPeaks.R
\name{sampleBackgroundPeaks}
\alias{sampleBackgroundPeaks}
\title{sampleBackgroundPeaks}
\usage{
sampleBackgroundPeaks(object, peak_set, niterations = 50)
}
\arguments{
\item{object}{fragmentCounts object}

\item{peak_set}{list of peak indices}

\item{niterations}{number of background sets to sample}
}
\value{
matrix with sampled counts for background sets
}
\description{
gets sample counts for background sets for a given peak set
}
\details{
input fragmentCounts object must include set of background peaks, computed
using \code{\link{getBackgroundPeakSets}}
}

