% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_synergy.R
\name{variability_synergy}
\alias{variability_synergy}
\title{variability_synergy}
\usage{
variability_synergy(object, annotations, background_peaks = NULL,
  variabilities = NULL, expectation = NULL, nbg = 25)
}
\arguments{
\item{object}{}

\item{annotations}{}

\item{background_peaks}{}

\item{variabilities}{}

\item{expectation}{}

\item{nbg}{}
}
\value{
synergy matrix
}
\description{
variability_synergy
}
\details{
should only be run on small number of motifs/kmers/peaksets (very slow!)
}

